Move StructureID enum from Common.h to Structure.h
